Add symbolic ex, new type param to Network.cs
Add a new type parameter to Network to allow for symbolics of the given type TS.
This does limit us to only networks where there is only one type of symbolic value, but this captures everything we've considered so far.
Also change code to use BigInteger instead of uint in most places: passing uints to the solver leads to responses where the value is chosen to lead to an integer overflow.