Many changes to Ast and type parameters
Change AstFunc to explicitly specify its return as a separate expr, and remove the Return Statement and Seq Statement, in favor of a List of Statement and this explicit Return.

Reorganize the Ast as AstAlt, which instead provides information about each node, matching the form of Batfish rather than the SRP model.

Add a BatfishBgpRoute struct to represent the routes generated from the Batfish configs.
As BatfishBgpRoute can define its own merge, we don't need to use one given to us by Ast anymore.

Add type parameters to Expr and its Evaluate method, which help deal with the types a bit more nicely than before. We still have an untyped State, but this hopefully won't be a problem in practice.