Re-add Seq and Return statements for arbitrary return pts
Change AstFunc's signature again to allow for multiple return points throughout the function, rather than a single one at the end.
Deprecate old Ast and testAst.json and use AstAlt and testAstAlternate.json instead.
Add a new method to Statement<T> to bind the result of a return expression
to a variable, for use when composing functions.