Add enough to Binder to parse testAst.json
Add cases to the Binder to enable us to parse testAst.json.
We'll want to think more about how we write out the strings describing our expressions and statements so that we can come up with a sensible approach for handling generic types.
Remove EdgesJsonConverter.cs now that we're using an alternative representation of the AST.