Change RoutingPolicies to use strings referencing Declarations
Change signature of RoutingPolicies to list the strings which define declarations we want for the policy.
This leads to some issues with trying to determine the classes of the declarations: effectively, we want a heterogeneous dictionary of AstFuncs with various choices of TArg and TResult, where we cast them appropriately at runtime and bail if that fails.