Refactor Statement and Expr to use a State with only a single type of stored variables.
Change State to only store one type Func<Zen<T>, Zen<T>>, and modify Statement and Expr to access it accordingly.
As in practice we only ever update the route and modify its state, this is safe for us to do to fix the problems caused by trying to cast between different types for the State.
There are still issues with some of the arithmetic expressions and likely others, but we can now actually get Evaluate to work for small examples as shown in AstFuncTests.cs.
Also add GetField and WithField classes for accessing the BatfishBgpRoute.