Attempt to address issues with serialization binder
Add tests module for SerializationBinder, and add a new function for determining the intended type for a string by breaking it apart into its components.
This has some issues when we want to try and determine the type for some statements, as it looks like the code is a bit inconsistent in referring to Route (BatfishBgpRoute) vs. a Pair of Bool and Route.
This needs to be cleaned up and rearranged so that we can be clear on what we want when.
Additionally, we need to make a concrete decision on what the types of Routes should be.
Assertions will want to have a notion of "does a route exist", which suggests storing a route explicitly with its boolean element.
However, transfer behavior does not yet map over the two components and handle them appropriately, and instead assumes a bare route as input (without being contained in a pair with a boolean) and we wind up needing quite a few additional AST terms to appropriately represent this.
We may want to add some shorthand if possible here, but that could be more trouble than it's worth: it may be best to instead just fix inconsistencies and let it be messy for now.