Reorganize AST code after refactor
Change NodeProperties to contain more AST information (decls and consts, for two).
Add a KaresansuiNode.cs class to store all this information when we want to write it out to a Network.
Change Ast to be generic over the T and TS arguments, and set those arguments in Program instead of Ast.
As a subsequent change, we should add modules to construct the different Asts we might want: for now, there is only one we need for our experiments so that can wait.
Use Symbolics in Ast.
Add new example instance to try out for sp4.