Update sp4, improve serializing
Add a feature to serializing to try and account for the problem that our temporal predicates want a different state from functions, meaning that we can't simply use the same AstBinder with a single type for all of them.
There may be a few ways to address this, but in any case we need a way to alternate depending on what is given to the serializer.
Requiring the user to list every type feels rather brutal, so a better approach would be to intelligently identify the desired state some other way.
As a trick to do so for now, add the option of putting a ! in front of the type for expressions, which enables them to be distinguished as "referring to time" or "not referring to time".
Otherwise, mess with Finally some more (still not working: might not be expressible without basically lifting the Expr<bool, T> to replace all references that require accessing a State<T> to accessing a State<Pair<T, BigInt>>.
Would there be any way to compile into predicates not referring to time and then add in time when we Evaluate?