Allow distinguishing type names for timed state vs untimed state.
Add features to serialization binder to distinguish timed and untimed state, and address a bug where the same type alias was used anytime a term recurred (since we were storing them all in a field, rather than constructing them freshly each time.
This fixes the bugs in the tests for AstSerializationBinderTests.cs.