Add TemporalOperators, Destination to Ast
Add TemporalOperators as a means to be more succinct in specifying predicates.
We can now express all predicates over routes in one place, and then add the timing
by using TemporalOperators.
Also change the AST to now also include the Destination with an IPAddress.
Note that since IPAddress is not serializable, we cannot currently serialize the AST anymore: we would need to modify the code to call ToString() first on Address.