Fix bug in BBR.Min, add current node's sol to inductive check State.cs
Add the model's state for the current node to the returned state for the inductive checks.
Also fix a bug in the code for BatfishBgpRouteExtensions.Min.
Previously, we had one-handed comparisons happening instead of two-way, which meant that we weren't considering the case that the other route was strictly better and returning immediately, but instead performing all the subsequent comparisons anyway.
Our tests didn't catch this because they only assessed the one direction, so more test coverage is needed.
I've made an attempt to fix CompareBy to do this now, but haven't changed the Min code yet: honestly, the Min code might be more readable just with all the Ifs.