Fix bug in FatTree constructor, re-add parallelism
Fix a bug in the FatTree constructor where safety and stable properties were not being combined properly to construct the modular and monolithic properties.
Re-add AsParallel() to CheckAnnotatedWith(), and switch the Dictionary for nodeTimes to a ConcurrentDictionary now that we're parallel.
Change reporting check times to happen on completing RunAnnotatedWithStats, regardless of exit point.