Change Hijack benchmark to use prefixes
Add a Destination field to BatfishBgpRoute to reason about routes
to particular prefixes.
Add a MinPrefix extension method to compare routes according to
what the intended routing destination is set to be.
There may be a problem with how this is written, as it currently
doesn't seem to work for a symbolic destination, but did for a concrete uint.
Add a MergeBy method to Lang to make it easier to write new merge
functions using old ones.
Add "Batfish" to the repo rider dictionary.

Fix a bug in FatTree.cs where names were being generated as PascalCase instead of lowercase.
Had to change ToString to ToLowerCaseString.