Simplify Benchmarks args and add more benchmarks
Remove timeout from benchmarks, and make destination optional: will be inferred for fattrees as the rightmost edge node.
Add alternate SpPathLength and ValleyFree benchmarks SpPathLengthWeak (a version without a safety property) and ValleyFreeLength (where we also require a short enough route).

Minor corrections and fixes to other benchmarks.
In particular, fix the bug in Hijack to allow a symbolic destination prefix to be used: makes for a janky constructor but fine for the time being!