Add a new Constants class for storing the AST's constants.
Constants will be any values set by the config and immutable during execution.
We can then look them up from Constants.

Currently, the idea would be to just list all of the possible types separately.
I think then we might be able to try and look up the right dictionary depending on the type trying to access the constant if we modify how Var is constructed.
There may be another clean way to do this just with casting everywhere.
Currently, this doesn't quite work since we need the constants to be defined by converting the serialized dictionary into the multiple dictionaries: this will probably need to be accomplished using a custom JsonConverter.