Clean up GenericMethod lookup, drop Bind
Get rid of Statement.Bind code, which didn't actually always do what it was intended for.
Instead, we just can convert to Funcs and then
compose at that stage.
Also fix the generic method invocation code, which allows us to fold With/GetField and First/Second into OpExprs, which is nice.
Note that an issue remains with deserializing generic exprs: because they take types as arguments, we either need to change serialization to give the types as fields, or change deserialization to supply those parameters.