Add new RouteEnvironment fields, deprecate Return
Deprecate Return in place of assigning to the input value and returning
it.
To do: add more of the RoutingPolicy logic to AstFunction so that it
properly handles the different cases when the result fields are set.
Add more RouteEnvironment fields to complete modelling of angler
environment type.
Delete outdated angler.json files.